enlarge unix_time field

+++ stream_context/src/main/scala/ConsumeKafka.scala
@@ -34,7 +34,7 @@ object ConsumeKafka {
         val lines = rdd.map(_._2)
         val ticksDF = lines.map( x => {
                                   val tokens = x.split(",")
-                                  Tick(tokens(0).toInt, tokens(1), tokens(2), tokens(3), tokens(4).toFloat, tokens(5).toInt)}).toDF()
+                                  Tick(tokens(0).toInt, tokens(1), tokens(2), tokens(3), tokens(4).toFloat, tokens(5).toLong)}).toDF()
         // val ticks_per_source_DF = ticksDF.groupBy("source_id")
         //                        .agg("quantity" -> "avg")//, "quantity" -> "sum")
         //                        .orderBy("source_id")
@@ -53,7 +53,7 @@ object ConsumeKafka {
 
  }
 }
-case class Tick(source_id: Int, item_sensed: String, subject_measured: String, sensor_location_name: String, quantity: Float, unix_time: Int)
+case class Tick(source_id: Int, item_sensed: String, subject_measured: String, sensor_location_name: String, quantity: Float, unix_time: Long)
 
 /** Lazily instantiated singleton instance of SQLContext */
